Twitter Analysis w/ Google Cloud and Spark - Kristin
I uploaded the .npz file for code that produced a matrix off of one twitter file, as well as a jupyter notebook with the code for the entire directory. After allowing the code to run for 5 hours with no output, I decided to stop it. I reviewed it and tried to optimize it, but it still did not complete the process within a reasonable amount of time. I would appreciate any feedback you could provide.